refactor: refactorizamos todas las llamdas a firebas en un solo archivo, en este commit lo hicimos con Get y Delete
