+++ src/Screens/HomeScreen.js
@@ -1,18 +1,11 @@
 import React, { useEffect, useLayoutEffect } from "react";
 import { TouchableOpacity, FlatList, View } from "react-native";
-import { Layout, themeColor, TopNav, Text } from "react-native-rapi-ui";
+import { themeColor } from "react-native-rapi-ui";
 import { FontAwesome5 } from "@expo/vector-icons";
 import { useNavigation } from "@react-navigation/native";
 
-import { database } from "../firebase/config";
-import {
-  collection,
-  doc,
-  onSnapshot,
-  orderBy,
-  query,
-} from "firebase/firestore";
 import TaskItem from "../Components/TaskItem";
+import { getTasks } from "../api/ApiFirebase";
 
 const HomeScreen = () => {
   const navigation = useNavigation();
@@ -20,20 +13,11 @@ const HomeScreen = () => {
   const [tasks, setTasks] = React.useState([]);
 
   useEffect(() => {
-    const dbRef = collection(database, "Tasks");
-    const q = query(dbRef, orderBy("createAt", "desc"));
-
-    const unsuscribe = onSnapshot(q, (querySnapshot) => {
-      setTasks(
-        querySnapshot.docs.map((x) => ({
-          id: x.id,
-          titleTask: x.data().titleTask,
-          descriptionTask: x.data().descriptionTask,
-          createAt: x.data().createAt,
-        }))
-      );
-    });
-    return unsuscribe;
+    try {
+      getTasks(setTasks);
+    } catch (e) {
+      alert(e);
+    }
   }, []);
 
   useLayoutEffect(() => {
